+++ src/frontend/noteUpdater.ts
