+++ src/midi/noteHandler.ts
